- minor codestyle refactoring in WorldImpl - generified BooleanMessage to TypeContainerMessage<Boolean>
