+++ src/main/java/project/websocket/messages/Message.java
@@ -4,7 +4,6 @@ import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonSubTypes.Type;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import project.websocket.messages.battle.*;
-import project.websocket.messages.bool.BooleanMessage;
 import project.websocket.messages.charlist.CharacterListRequestMessage;
 import project.websocket.messages.charlist.CharacterListResponseMessage;
 import project.websocket.messages.charlist.CreateCharacterRequestMessage;
@@ -12,6 +11,7 @@ import project.websocket.messages.charlist.DeleteCharacterRequestMessage;
 import project.websocket.messages.matchmaking.LobbyConfirmationMessage;
 import project.websocket.messages.matchmaking.LobbyRequestMessage;
 import project.websocket.messages.matchmaking.MatchmakingNotificationMessage;
+import project.websocket.messages.typecontainer.BooleanMessage;
 
 @SuppressWarnings("unused")
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.PROPERTY, property = "class")
