+++ src/main/java/project/websocket/messages/typecontainer/TypeContainerMessage.java
@@ -1,0 +1,19 @@
+package project.websocket.messages.typecontainer;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import project.websocket.messages.Message;
+
+import javax.validation.constraints.NotNull;
+
+public class TypeContainerMessage<T> extends Message {
+    private final T value;
+
+    public TypeContainerMessage(@JsonProperty ("value")
+                                @NotNull T value) {
+        this.value = value;
+    }
+
+    public T getValue() {
+        return value;
+    }
+}
